requirements.txt
allure_python_commons==2.13.2
playwright==1.37.0
pytest==7.4.0
pytest_playwright==0.4.2
python-slugify==8.0.1
./pytest_playwright_allure-0.0.1-py3-none-any.whl